# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartLeads API Documentation",
    "version": "1.0.0",
    "description": "Comprehensive REST API specification for the SmartLeads Lead Management Dashboard. Built with Node.js, Express, TypeScript, and MongoDB.",
    "contact": {
      "name": "Rahul Kumar",
      "url": "https://github.com/rahulkrs9142/smart-leads"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "https://smart-leads-backend.onrender.com",
      "description": "Render Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
...
  },
  "paths": {
...
    "/api/leads": {
      "get": {
        "summary": "List and filter leads",
        "description": "Supports pagination, status filtering, source filtering, and text search across lead name and email",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["New", "Contacted", "Qualified", "Lost"] } },
          { "name": "source", "in": "query", "schema": { "type": "string", "enum": ["Website", "Instagram", "Referral"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search keyword for name or email" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["latest", "oldest"] } }
        ],
        "responses": {
          "200": { "description": "Leads retrieved with pagination metadata" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new lead",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "source"],
                "properties": {
                  "name": { "type": "string", "example": "Tech Corp Lead" },
                  "email": { "type": "string", "format": "email", "example": "lead@techcorp.com" },
                  "source": { "type": "string", "enum": ["Website", "Instagram", "Referral"], "example": "Website" },
                  "status": { "type": "string", "enum": ["New", "Contacted", "Qualified", "Lost"], "default": "New" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Lead created successfully" },
          "400": { "description": "Validation error" }
        }
      }
    },
...
  }
}